convert connection_to to long since it used in timer code

+++ keepalived/include/check_api.h
@@ -31,7 +31,7 @@
 typedef struct _conn_opts {
 	struct sockaddr_storage		dst;
 	struct sockaddr_storage		bindto;
-	unsigned int			connection_to; /* connection time-out */
+	unsigned long			connection_to; /* connection time-out */
 #ifdef _WITH_SO_MARK_
 	unsigned int			fwmark; /* to mark packets going out of the socket using SO_MARK */
 #endif
